Extract DevPtrCast to device_ptr_cast.h
